refactor: pieces 패키지 분리
