+++ ReadmeDir/Mission2.md
@@ -17,12 +17,12 @@
   - [X] 추가한 Pawn이 정상적으로 보드에 추가되었는지 확인하는 테스트 추가
 
 ## Refactor
-- [ ] 패키지 분리 
-  - [ ] pieces 패키지 생성 
-  - [ ] Pawn 클래스와 PawnTest 클래스를 pieces 패키지로 이동 
-  - [ ] 컴파일 문제가 발생하면 해결하고 테스트 코드를 실행해 모두 성공하는지 확인
-- [ ] 추가 테스트 
-  - [ ] 다른 객체를 추가하면 컴파일 에러가 발생하는지 확인하는 테스트 추가
+- [X] 패키지 분리 
+  - [X] pieces 패키지 생성 
+  - [X] Pawn 클래스와 PawnTest 클래스를 pieces 패키지로 이동 
+  - [X] 컴파일 문제가 발생하면 해결하고 테스트 코드를 실행해 모두 성공하는지 확인
+- [X] 추가 테스트 
+  - [X] 다른 객체를 추가하면 컴파일 에러가 발생하는지 확인하는 테스트 추가
 - [ ] 중복 제거 
   - [ ] 코드에서 중복된 부분 확인 및 제거 
   - [ ] @Before 애노테이션을 활용해 중복 제거
