+++ src/main/java/softeer2nd/chess/Board.java
@@ -1,5 +1,7 @@
 package softeer2nd.chess;
 
+import softeer2nd.chess.pieces.Pawn;
+
 import java.util.ArrayList;
 import java.util.List;
