+++ src/test/java/softeer2nd/chess/pieces/PawnTest.java
@@ -1,7 +1,8 @@
-package softeer2nd.chess;
+package softeer2nd.chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import softeer2nd.chess.pieces.Pawn;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
